OGM-1006 Enable the tests for the mass-indexer with Neo4j embedded
