+++ integrationtest/src/test/java/org/hibernate/ogm/test/integration/neo4j/embedded/EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT.java
@@ -6,14 +6,10 @@
  */
 package org.hibernate.ogm.test.integration.neo4j.embedded;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-
 import org.hibernate.ogm.datastore.neo4j.Neo4j;
 import org.hibernate.ogm.datastore.neo4j.Neo4jProperties;
 import org.hibernate.ogm.jpa.HibernateOgmPersistence;
-import org.hibernate.ogm.test.integration.testcase.ModuleMemberRegistrationScenario;
-import org.hibernate.ogm.test.integration.testcase.model.Member;
+import org.hibernate.ogm.test.integration.neo4j.Neo4jModuleMemberRegistrationScenario;
 import org.hibernate.ogm.test.integration.testcase.util.ModuleMemberRegistrationDeployment;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
@@ -23,7 +19,6 @@ import org.jboss.shrinkwrap.descriptor.api.persistence20.PersistenceDescriptor;
 import org.jboss.shrinkwrap.descriptor.api.persistence20.PersistenceUnit;
 import org.jboss.shrinkwrap.descriptor.api.persistence20.PersistenceUnitTransactionType;
 import org.jboss.shrinkwrap.descriptor.api.persistence20.Properties;
-import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
@@ -33,12 +28,13 @@ import org.junit.runner.RunWith;
  * @author Guillaume Scheibel &lt;guillaume.scheibel@gmail.com&gt;
  */
 @RunWith(Arquillian.class)
-public class EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT extends ModuleMemberRegistrationScenario {
+public class EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT extends Neo4jModuleMemberRegistrationScenario {
 
 	@Deployment
 	public static Archive<?> createTestArchive() throws Exception {
 		return new ModuleMemberRegistrationDeployment
 				.Builder( EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT.class )
+					.addClasses( Neo4jModuleMemberRegistrationScenario.class )
 				.persistenceXml( persistenceXml() )
 				.manifestDependencies( "org.hibernate.ogm:${hibernate-ogm.module.slot} services, org.hibernate.ogm.neo4j:${hibernate-ogm.module.slot} services" )
 				.createDeployment();
@@ -54,30 +50,10 @@ public class EmbeddedNeo4jResourceLocalModuleMemberRegistrationIT extends Module
 				.getOrCreateProperties();
 		PersistenceDescriptor persistenceDescriptor = propertiesContext
 				.createProperty().name( Neo4jProperties.DATASTORE_PROVIDER ).value( Neo4j.EMBEDDED_DATASTORE_PROVIDER_NAME ).up()
-				.createProperty()
-				.name( Neo4jProperties.DATABASE_PATH )
-				.value( EmbeddedNeo4jJtaModuleMemberRegistrationIT.neo4jFolder() )
-				.up()
+				.createProperty().name( Neo4jProperties.DATABASE_PATH ).value( EmbeddedNeo4jJtaModuleMemberRegistrationIT.neo4jFolder() ).up()
 				.createProperty().name( "hibernate.search.default.directory_provider" ).value( "ram" ).up()
 				.createProperty().name( "wildfly.jpa.hibernate.search.module" ).value( "org.hibernate.search.orm:${hibernate-search.module.slot}" ).up()
 				.up().up();
 		return persistenceDescriptor;
-	}
-
-	@Override
-	public void shouldBeAbleToFindMemberByEmail() throws Exception {
-	}
-
-	@Test
-	public void shouldFindPersistedMemberByIdWithNativeQuery() throws Exception {
-		Member newMember = memberRegistration.getNewMember();
-		newMember.setName( "Giovanni Doe" );
-		memberRegistration.register();
-
-		String nativeQuery = "MATCH (n:Member {id: " + newMember.getId() + "}) RETURN n";
-		Member found = memberRegistration.findWithNativeQuery( nativeQuery );
-
-		assertNotNull( "Expected at least one result using a native query", found );
-		assertEquals( "Native query hasn't found a new member", newMember.getName(), found.getName() );
 	}
 }
